Handle session counter overflow.
While it should be possible to negotiate a session within usize::MAX
attempts we nevertheless have to handle the theoretical possibility that
the counter which monotonically increases with every new session state
added does in fact overflow. The remedy is somewhat brutal―we remove all
session states and reset the counter to 0.