Sign prekeys.
Sign the ephemeral public key in prekey bundles with the private identity key.
Conversely, when initiating a session from a prekey bundle, verify the
signature.

While this weakens deniability in one direction, ensuring that the ephemeral
keys in prekey bundles are authentic prevents abuse scenarios with forged
prekey bundles where the ephemeral key is exchanged with the goal of recording
ciphertext that can later be decrypted if the long-term private key is
compromised.

Especially in multi-device scenarios with mechanisms for extending trust from
one device to another and thus from one identity key to another (e.g. through
cloning identity keys or (cross-)signing of identity keys) it is otherwise
possible for a malicious device or server to forge prekey bundles with their
own ephemeral key pair and another device's identity. The goal thereby is to
make the victim's device trust the malicious device through the existing trust on
an identity key.

If successful, the malicious device will thus receive (and can record) all the
ciphertext sent by the compromised client in conversations that the malicious
device participates in. If it now ever manages to compromise the long-term
private key that corresponds to the public key used in the attack, it can
initialise its session and decrypt all of the recorded ciphertext.

This problem does not affect prekey messages, because here the attacker has to
initialise his session first, so even though he can try to send a forged prekey
message with someone else's public identity key, the recipient will not be able
to decrypt it (triple-DH will yield a different secret) and thus fail to
initialise the session.