`InvalidMessage` -> `InvalidSignature`.
Use `InvalidSignature` in `try_skipped_message_keys` if signature
validation fails. This is consistent with other instances where the
signature is checked. The case where the message counter denotes
wrong message keys is also not different to a failed signature check
in `SessionState::decrypt` where the message counter equals the chain
key index.