Change decryption order in `Session::decrypt`.
Previously, when receiving a prekey message we always tried to look up
the corresponding prekey first, before attempting to decrypt the
message. This can be inefficient (in case the current session is
actually the correct one) and cause errors (in case the prekey ID range
has been exhausted and a new but different prekey exists for that ID).

If only Alice initialises a session and Bob never writes back to Alice,
she will continue to send prekey messages which may eventually trigger
such an error on Bob's side.

With this commit we reverse the order and first try to simply decrypt
the message and in case the signature does not match or no session state
could be found we fall back on trying to initialise a new session state
with the prekey mentioned in the prekey message. Other errors (e.g.
duplicate message et al.) are simply returned directly.